pyScripts/requirements.txt
matplotlib==3.0.0
six==1.11.0
pandas==0.23.4
Keras==2.2.4
numpy==1.16.3+mkl
Orange==2.7.8
python_dateutil==2.8.0
heatmap==2.2.1
scikit_learn==0.20.3
